+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.5](../../tags/v2.1.5) - 2023-08-05
+### Changed
+- Collapse excess dots
+
 ## [2.1.4](../../tags/v2.1.4) - 2023-08-05
 ### Changed
 - Use regex unicode categories and specify ranges more precisely